Devoir 4/requirements.txt
setuptools
streamlit
requests
matplotlib
pandas
torch
torchvision
pillow
validators
gunicorn
flask
flask_cors
flask_sslify